Renamed file isImage.js to isImageObj.js
